staging: prima: switch back to LA.BR.1.2.9_rb1.14 stack
@As007 Runs into an issue with the latest prima stack of CAF
LA.BR.1.2.9_rb1.15, That is hard reboot on wifi termination.
So after examining the logs i found that the latest prima release is
failing with null pointer allocations on an average a lot of garbage
packets are bundling up and crashing the kernel on AP termination
requests and causing hard reboots. #Blame_CAF

STEPS TO REPRODUCE :-
1:- Start wifi connection and connect to a wifi network and browse
atleast for 2-3 mins
2:- Try turning of wifi it will give you a screen freeze and Hard reboot
3:- Or simply you could try by switching on Airplane mode and get the
issue live

Workarounds:- Use older prima stack for now

This reverts commit 617a6b98deeefa21192d47ae5e98016de07eba5a.